eventInitDict parameter to GamepadEvent constructor should be optional https://bugs.webkit.org/show_bug.cgi?id=177308
Reviewed by Sam Weinig.

Source/WebCore:

eventInitDict parameter to GamepadEvent constructor should be optional to match Firefox
and Chrome. Having such dictionary parameter is also not valid Web IDL:
"""
If the type of an argument is a dictionary type (or a union type that has a dictionary as one
of its flattened member types), and that dictionary type and its ancestors have no required
members, and the argument is either the final argument or is followed only by optional arguments,
then the argument must be specified as optional. Such arguments are always considered to have a
default value of an empty dictionary, unless otherwise specified.
"""

Test: gamepad/gamepad-event.html

* Modules/gamepad/GamepadEvent.idl:

LayoutTests:

Add layout test coverage.

* gamepad/gamepad-event-expected.txt: Added.
* gamepad/gamepad-event.html: Added.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@222348 268f45cc-cd09-0410-ab3c-d52691b4dbfc