Layout Test media/modern-media-controls/scrubber-support/scrubber-support-click.html is flaky https://bugs.webkit.org/show_bug.cgi?id=177930
Patch by Antoine Quint <graouts@apple.com> on 2017-10-05
Reviewed by Dean Jackson.

In cases where this test failed, it actually worked as expected but the "timeupdate" event would be fired
twice, most likely because it was registered twice. We now use an asynchronous expectation to run the test
which should only register the event listener once and make the test more reliable.

* media/modern-media-controls/scrubber-support/scrubber-support-click-expected.txt:
* media/modern-media-controls/scrubber-support/scrubber-support-click.html:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@222923 268f45cc-cd09-0410-ab3c-d52691b4dbfc