It should not be possible to submit a form that is disconnected https://bugs.webkit.org/show_bug.cgi?id=178099
Reviewed by Sam Weinig.

LayoutTests/imported/w3c:

Rebaseline several WPT tests now that more checks are passing.

* web-platform-tests/dom/events/Event-dispatch-click-expected.txt:
* web-platform-tests/html/semantics/forms/the-button-element/button-click-submits-expected.txt:

Source/WebCore:

It should not be possible to submit a form that is disconnected. Both Firefox and Chrome agree with the specification.

This is as per:
https://html.spec.whatwg.org/multipage/form-control-infrastructure.html#form-submission-algorithm (step 1)
which refers to:
https://html.spec.whatwg.org/multipage/links.html#cannot-navigate

Form cannot navigate when it is disconnected.

No new tests, rebaselined existing tests.

* html/HTMLFormElement.cpp:
(WebCore::HTMLFormElement::prepareForSubmission):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@223117 268f45cc-cd09-0410-ab3c-d52691b4dbfc