[JSC] modules can be visited more than once when resolving bindings through "star" exports as long as the exportName is different each time https://bugs.webkit.org/show_bug.cgi?id=178308
Reviewed by Mark Lam.

JSTests:

* test262.yaml:

Source/JavaScriptCore:

With the change of the spec[1], we now do not need to remember star resolution modules.
We reflect this change to our implementation. Since this change is covered by test262,
this patch improves the score of test262.

We also add logging to ResolveExport to debug it easily.

[1]: https://github.com/tc39/ecma262/commit/a865e778ff0fc60e26e3e1c589635103710766a1

* runtime/AbstractModuleRecord.cpp:
(JSC::AbstractModuleRecord::ResolveQuery::dump const):
(JSC::AbstractModuleRecord::resolveExportImpl):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@223894 268f45cc-cd09-0410-ab3c-d52691b4dbfc