Web Inspector: Layer mutations should be purely based on layerId, not based on nodeId https://bugs.webkit.org/show_bug.cgi?id=178554
Patch by Ross Kirsling <ross.kirsling@sony.com> on 2017-10-24
Reviewed by Devin Rousso.

Source/WebInspectorUI:

* UserInterface/Controllers/LayerTreeManager.js:
(WI.LayerTreeManager.prototype.layerTreeMutations):
Looking for special cases involving nodeIds is incorrect, as nodeIds need not be unique in the layer list (such
as when an element and a pseudo-element thereof each give rise to a layer). A layer object marked "preserved" in
this way shares no data with its predecessor, meaning that no consumer can act upon this so-called preservation.
A preserved layer should be nothing more or less than a recycled layerId (the thing that *is* unique).

LayoutTests:

* inspector/layers/layer-tree-manager-expected.txt:
* inspector/layers/layer-tree-manager.html:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@223914 268f45cc-cd09-0410-ab3c-d52691b4dbfc