[GTK] Crash in WebCore::SelectionRangeData::apply https://bugs.webkit.org/show_bug.cgi?id=177808
Reviewed by Carlos Garcia Campos.

The problem appears to be a name collision between WebCore::SelectionData and
WebCore::SelectionData. Yes, those are the same names. It seems the destructor for a
Source/WebCore/platform/gtk/SelectionData.h SelectionData is being called on a
Source/WebCore/rendering/SelectionRangeData.cpp SelectionData. Yikes. Can you forgive the
poor linker for not being magic?

This fixes crashes in approximately one bajillion tests.

* rendering/SelectionRangeData.cpp: Move SelectionData into an anonymous namespace.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@224087 268f45cc-cd09-0410-ab3c-d52691b4dbfc