ASSERTION FAILED: internalValuesConsistent(m_url) in WebCore::URLParser::URLParser https://bugs.webkit.org/show_bug.cgi?id=178861
Reviewed by Tim Horton.

Source/WebCore:

This is a dark corner of the URL spec that has wildly different behavior in different browsers.
The assertion fired when we have a file URL with a ? or a # after file:// and it was because
m_pathAfterLastSlash was still 0. We definitely shouldn't assert. I'm making us consistent with
other cases where we have nothing there and add an implied missing slash.
        
Covered by new API tests.

* platform/URLParser.cpp:
(WebCore::URLParser::parse):

Tools:

* TestWebKitAPI/Tests/WebCore/URLParser.cpp:
(TestWebKitAPI::TEST_F):



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@224202 268f45cc-cd09-0410-ab3c-d52691b4dbfc