ServiceWorkerContainer should take a PendingActivity while a service worker job is pending https://bugs.webkit.org/show_bug.cgi?id=179321
Reviewed by Youenn Fablet.

ServiceWorkerContainer should take a PendingActivity while a service worker job is pending.
ServiceWorkerContainer is an ActiveDOMObject and taking such pending activity makes sure
the object stays alive while a job is pending (and a promise is not resolved yet). It also
makes sure the document does not get suspended in the middle of a job.

* workers/service/ServiceWorkerContainer.cpp:
(WebCore::ServiceWorkerContainer::scheduleJob):
(WebCore::ServiceWorkerContainer::jobDidFinish):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@224496 268f45cc-cd09-0410-ab3c-d52691b4dbfc