REGRESSION(r224179): [GTK] Several WebViewEditor tests are failing since r224179 https://bugs.webkit.org/show_bug.cgi?id=179366
Reviewed by Michael Catanzaro.

Source/WebKit:

In r224179, webkit_web_view_can_execute_editing_command() was optimized to use the sync path for commands
supported by the WebViewEditorState, but the state requires a redraw to be up to date. We can't know if
WebViewEditorState is in sync, when webkit_web_view_can_execute_editing_command() is called, so we always need
to ask the web process.

* UIProcess/API/glib/WebKitWebView.cpp:
(webkit_web_view_can_execute_editing_command):

Tools:

Use always a toplevel window for WebViewEditor tests instead of creating and destroying one only to flush editor
state.

* TestWebKitAPI/Tests/WebKitGtk/TestWebViewEditor.cpp:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@224567 268f45cc-cd09-0410-ab3c-d52691b4dbfc