[ios-simulator] API test WebKit.BundleParameters is a flaky failure https://bugs.webkit.org/show_bug.cgi?id=178363 <rdar://problem/35043144>
Reviewed by Brent Fulgham.

* wtf/RefCounter.h:
(WTF::RefCounter<T>::Count::deref):
(WTF::RefCounter<T>::Count::refCounterWasDeleted):
(WTF::RefCounter<T>::~RefCounter):
If a RefCounter::Count is deref()'d and drops its RefCounter's value to
zero, and the RefCounter is deleted in that valueDidChangeCallback, it
will delete the Counter that is in the middle of deref(). Keep track
of whether we're inside the callback and defer the deletion until
the end of deref().

* TestWebKitAPI/Tests/WTF/RefCounter.cpp:
(TestWebKitAPI::TEST):
Add a test.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@224597 268f45cc-cd09-0410-ab3c-d52691b4dbfc