Compositing layers in UIWebView are often missing content after scrolling https://bugs.webkit.org/show_bug.cgi?id=179564 rdar://problem/33885639
Reviewed by Tim Horton.

When dropping backing store for layers outside the viewport, we rely on a compositing
layer flush to compute layer visibility and bring the backing store back when necessary.

In UIWebView, we have a mechanism to trigger flushing on user scrolling but it only triggers
when there are tiled layers. Remove that condition so it triggers always, now that
backing store availability also depends on this.

* page/ios/FrameIOS.mm:
(WebCore::Frame::viewportOffsetChanged):
(WebCore::anyFrameHasTiledLayers): Deleted.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@224716 268f45cc-cd09-0410-ab3c-d52691b4dbfc