Consolidate mach-lookup calls in WebProcess sandbox https://bugs.webkit.org/show_bug.cgi?id=179650 <rdar://problem/35367154>
Reviewed by Dean Jackson.

Consolidate the various calls to 'allow mach-lookup' imported during Bug 179548 into the main
function in the sandbox profile. Remove some entries that are not needed by WebKit.

Remove the statement to grant global permission to perform mach-lookup for any xpc
service name that was copied into this sandbox profile in Bug 179548. We started blocking
this in macOS 10.13 and want to continue to do so.
        
The earlier "grant global permission" in 'system.sb' apparently allowed some xpc lookups to occur
before we hit the block declaration in the WebContent sandbox. Now that we are consistently blocking
mach lookups we need to add whitelist entries for a few more services to avoid creating new
sandbox violations.

* WebProcess/com.apple.WebProcess.sb.in:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@224821 268f45cc-cd09-0410-ab3c-d52691b4dbfc