REGRESSION (r224592): oss-fuzz: jsc: Null-dereference READ in JSC::JSCell::isObject (4216) https://bugs.webkit.org/show_bug.cgi?id=179763 <rdar://problem/35550513>
Reviewed by Keith Miller.

JSTests:

Just adding a slightly cleaned-up version of the original fuzzer-found test.

* stress/tdz-this-in-try-catch.js: Added.
(__v_6388):
(__v_6392):

Source/JavaScriptCore:

Fix null pointer dereference caused by an eliminated tdz_check

The problem was when doing an OSR entry in DFG while |this| was null
(because super() had not yet been called in the constructor of this
subclass), it would be marked as non-null, and the tdz_check eliminated.

* dfg/DFGInPlaceAbstractState.cpp:
(JSC::DFG::InPlaceAbstractState::initialize):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@224915 268f45cc-cd09-0410-ab3c-d52691b4dbfc