Fix occasional crash when adding and removing videos https://bugs.webkit.org/show_bug.cgi?id=179792
Reviewed by Geoffrey Garen.

In taskCompleted we null out session, and in resume we call _restart on the main thread.
If _restart is called after taskCompleted, we want to do nothing.
Right now we are calling a method on self.session.loader which is a PlatformMediaResourceLoader&
but in ObjC if session is null it will call PlatformMediaResourceLoader::requestResource with a
null this pointer, which crashes.  Let's not crash.

* platform/network/cocoa/WebCoreNSURLSession.mm:
(-[WebCoreNSURLSessionDataTask _restart]):
Early return if we don't have a session.



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@224960 268f45cc-cd09-0410-ab3c-d52691b4dbfc