[WTF] Use system endianess information instead of relying CPU information https://bugs.webkit.org/show_bug.cgi?id=179861
Reviewed by JF Bastien.

Currently, we use known CPU information to determine endianess of the current
architecture. But this means that unknown CPU does not work well in WebKit
withou modifying code. Instead, we use compiler or system's macro to determine
endianess. This paves the way to dropping many CPUs in Platform.h by replacing
them with CPU(UNKNOWN)[1].

[1]: https://bugs.webkit.org/show_bug.cgi?id=179243

* wtf/Platform.h:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@225032 268f45cc-cd09-0410-ab3c-d52691b4dbfc