[Web Animations] Adopt KeyframeList in KeyframeEffect https://bugs.webkit.org/show_bug.cgi?id=179941 <rdar://problem/35666924>
Reviewed by Darin Adler.

KeyframeEffect used to model its keyframes by creating a new Keyframe struct and storing them in a Vector<Keyframe>.
However, there already is a way to model keyframes in WebCore using the KeyframeList class. We now adopt this class
in KeyframeEffect, which makes it possible to run hardware-composited animations using RenderBoxModelObject::startAnimation()
in a follow-up patch, since this function expects a KeyframeList.

No news tests, this is an internals change and does not yield any changes in Web content.

* animation/KeyframeEffect.cpp:
(WebCore::KeyframeEffect::KeyframeEffect): Create a default m_keyframes at construction time.
(WebCore::KeyframeEffect::processKeyframes): Create a temporary KeyframeList as we parse keyframes passed from the JS API
and assign it to m_keyframes upon successful parsing. We also use cssPropertyId to refer to a CSSPropertyID to make the code
clearer and ensure there is no confusion with JS properties that we manipulate in this function.
(WebCore::KeyframeEffect::computeStackingContextImpact): Use the list of CSS properties exposed on KeyframeList directly rather
than enumerating each keyframe and use cssPropertyId to refer to individual CSSPropertyID values for consistency.
(WebCore::KeyframeEffect::applyAtLocalTime): Use the list of CSS properties exposed on KeyframeList directly rather than
using the properties specified on the first keyframe only. We also remove a stale FIXME comment which was addressed in r224957.
* animation/KeyframeEffect.h: Remove the Keyframe struct and change the type of m_keyframes from Vector<Keyframe> to KeyframeList.
* rendering/style/KeyframeList.h:
(WebCore::KeyframeList::KeyframeList): We remove the default insertions at KeyframeList construction time since it requires
an immediate call to clear() when doing any meaningful work with a KeyframeList. We also add the default move assignment operator
to allow KeyframeEffect::processKeyframes() to make a move assignment from the temporary keyframe list we build at parsing to the
member variable.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@225099 268f45cc-cd09-0410-ab3c-d52691b4dbfc