RenderBlockFlow::layoutRunsAndFloatsInRange is O(n^2) for runs of inlines without any text https://bugs.webkit.org/show_bug.cgi?id=179950
Reviewed by Simon Fraser.

It calls createBidiRunsForLine for each line. createBidiRunsForLine traverses past the end of the line until
it finds the end of the current bidi run. If there is no text in the flow, it never finds anything and traverses
the entire flow. This is O(n^2) for the number of renderers in the flow.

Tested by PerformanceTests/Layout/inline-layout-no-text.html

* platform/text/BidiResolver.h:
(WebCore::BidiResolverBase::needsContinuePastEnd const):
(WebCore::BidiResolverBase::needsContinuePastEndInternal const):
(WebCore::DerivedClass>::createBidiRunsForLine):

    When past end of line call needsContinuePastEnd() to see if we need to continue searching for the end of the bidi run.

* rendering/InlineIterator.h:
(WebCore::InlineBidiResolver::needsContinuePastEndInternal const):

    InlineBidiResolver only returns runs up to the last renderer on the line and can bail out.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@225110 268f45cc-cd09-0410-ab3c-d52691b4dbfc