Clean up spanners before creating nested column context https://bugs.webkit.org/show_bug.cgi?id=180107 <rdar://problem/35686655>
Reviewed by Antti Koivisto.

Source/WebCore:

When an existing spanner placeholder is moved into a newly constructed (and nested)
multicolumn context, we figure it's not valid anymore and end up destroying it
(see RenderMultiColumnFlow::fragmentedFlowDescendantInserted).
This is very unfortunate since as we climb back on the stack, we could hit this renderer as
the newly inserted child.

This patch proactively removes the invalid placeholders and moves the associated spanners back to their
original position.

Test: fast/multicol/crash-when-constructing-nested-columns.html

* rendering/RenderMultiColumnFlow.h:
* style/RenderTreeUpdaterMultiColumn.cpp:
(WebCore::RenderTreeUpdater::MultiColumn::createFragmentedFlow):
RenderTreeUpdater::MultiColumn::destroyFragmentedFlow still relies on the placeholder removal
logic in RenderMultiColumnFlow::fragmentedFlowDescendantInserted.

LayoutTests:

* fast/multicol/crash-when-constructing-nested-columns-expected.txt: Added.
* fast/multicol/crash-when-constructing-nested-columns.html: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@225255 268f45cc-cd09-0410-ab3c-d52691b4dbfc