Flaky crash in WebCore::DOMGuardedObject::clear() during service worker tests https://bugs.webkit.org/show_bug.cgi?id=180045 <rdar://problem/35737288>
Patch by Youenn Fablet <youenn@apple.com> on 2017-11-29
Reviewed by Chris Dumez.

Manually tested by running concurrently service worker tests using FetchEvents which store promise references.

Before the patch, on workers, clearing of DOMGuardedObjects happens at the time WorkerGlobalScope is destroyed.
This is too late as it is expected that the JSDOMGlobalObject is still alive.

This patch adds a clearDOMGuardedObjects method on JSWorkerGlobalScopeBase.
It is called when stopping a WorkerThread, just before releasing the strong reference to JSWorkerGlobalScopeBase.

* bindings/js/JSDOMGuardedObject.h:
* bindings/js/JSWorkerGlobalScopeBase.cpp:
(WebCore::JSWorkerGlobalScopeBase::clearDOMGuardedObjects):
* bindings/js/JSWorkerGlobalScopeBase.h:
* bindings/js/WorkerScriptController.cpp:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@225291 268f45cc-cd09-0410-ab3c-d52691b4dbfc