WKAttachmentTests.InPlaceImageAttachmentToggleDisplayMode times out on macOS bots https://bugs.webkit.org/show_bug.cgi?id=180200
Reviewed by Alexey Proskuryakov.

Rewrites an API test to avoid checking against the hard-coded platform-dependent size of an attachment element.
Instead, first insert the attachment element as an icon and compute its size, and then check that the size of
the element is restored to this original value after toggling the display mode to in-place mode and back.

* TestWebKitAPI/Tests/WebKitCocoa/WKAttachmentTests.mm:
(-[TestWKWebView attachmentElementSize]):
(-[TestWKWebView waitForAttachmentElementSizeToBecome:]):
(TestWebKitAPI::TEST):
(platformAttachmentIconElementSize): Deleted.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@225332 268f45cc-cd09-0410-ab3c-d52691b4dbfc