Cursor is not visible after exiting full screen video https://bugs.webkit.org/show_bug.cgi?id=180247 <rdar://problem/33885922>
Patch by Antoine Quint <graouts@apple.com> on 2017-12-04
Reviewed by Dean Jackson.

Source/WebCore:

There is a dedicated NSCursor method to temporarily hide the mouse cursor while the mouse is idle,
so we use this platform functionality, already exposed through the page chrome, instead which
implements the expected behavior. Now, the mouse cursor is hidden while the mouse is stationary
when a <video> is fullscreen, as before, but as soon as the user exits fullscreen, the mouse cursor
reappears.

No test provided as I don't believe this platform behavior can be tested, there is no API to query
whether the cursor is visible.

* page/EventHandler.cpp:
(WebCore::EventHandler::cancelAutoHideCursorTimer):
(WebCore::EventHandler::autoHideCursorTimerFired):

LayoutTests:

Remove the existing test which would query a cursor state that is no longer relevant.

* fullscreen/video-cursor-auto-hide-expected.txt: Removed.
* fullscreen/video-cursor-auto-hide.html: Removed.
* platform/gtk/TestExpectations:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@225494 268f45cc-cd09-0410-ab3c-d52691b4dbfc