WebServiceWorkerProvider should use Cancellation error to notify DTL that it cannot handle a fetch https://bugs.webkit.org/show_bug.cgi?id=180584
Patch by Youenn Fablet <youenn@apple.com> on 2017-12-08
Reviewed by Alex Christensen.

Source/WebCore:

Previously, for each cross origin fetch that is going through a service worker and service worker is not handling the fetch,
we return an AccessControl error so that DocumentThreadableLoader will do preflight and regular load through the network.
This error is wrongly logged in the Inspector.
Change error type to Cancellation so that the Inspector does not log it.

* loader/DocumentThreadableLoader.cpp:
(WebCore::DocumentThreadableLoader::didFail):

Source/WebKit:

* WebProcess/Network/WebLoaderStrategy.cpp:
(WebKit::WebLoaderStrategy::scheduleLoad):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@225705 268f45cc-cd09-0410-ab3c-d52691b4dbfc