Unreviewed webkitpy fix on Linux platforms after r225721.
Add an empty 'linux' entry to the PUBLIC_TABLE mapping in the
VersionNameMap class. This avoids asserting in this class's
mapping_for_platform() method when the code is evaluated on a Linux
system, which happens when the static VERSION_FALLBACK_ORDER member in
the WinCairoPort class is constructed. This can occur during a
test-webkitpy run or when trying to construct a WPEPort instance in
PortFactory, which isn't possible because again the WinCairoPort
instance is constructed first, leading to the same assert in both cases.

* Scripts/webkitpy/common/version_name_map.py:
(VersionNameMap.__init__):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@225738 268f45cc-cd09-0410-ab3c-d52691b4dbfc