ononline, onoffline and navigator.onLine don't work if Parallels or VMWare is installed https://bugs.webkit.org/show_bug.cgi?id=32327
Reviewed by Alexey Proskuryakov.

When determining if we are online, ignore virtual interfaces added on the host machine
by Parallels and VMWare. This is needed because those interfaces are always up, whether
or not the virtual machine is running. This was causing navigator.onLine to always return
true on the host machine when Parallels or VMWare was installed.

Note that it is safe to ignore these interfaces and that it does not cause issues when
running Safari inside the virtual machine because those virtual interfaces are only
exposed on the host machine. Inside, the virtual machine, we see the usual en0 interface.

* platform/network/mac/NetworkStateNotifierMac.cpp:
(WebCore::NetworkStateNotifier::updateStateWithoutNotifying):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@226238 268f45cc-cd09-0410-ab3c-d52691b4dbfc