REGRESSION (r224301?): LayoutTest http/tests/workers/service/registration-task-queue-scheduling-1.html is a flaky failure https://bugs.webkit.org/show_bug.cgi?id=179342 <rdar://problem/35478161>
Reviewed by Youenn Fablet.

* TestExpectations:
Unmark the test as flaky.

* http/tests/workers/service/registration-task-queue-scheduling-1-expected.txt:
* http/tests/workers/service/resources/registration-task-queue-scheduling-1-second-window.html:
* http/tests/workers/service/resources/registration-task-queue-scheduling-1.js:
Update test to register a scriptURL that actually fails. When registration succeeds, the test
logs lines indicating that registration succeeded unexpectedly. Also, when registration
succeeds, the registration of the top frame are no longer guaranteed to be resolved before
the ones of the subframe (because it involves things like network loads).


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@226722 268f45cc-cd09-0410-ab3c-d52691b4dbfc