Each variant of a polymorphic inlined call should be exitOK at the top of the block https://bugs.webkit.org/show_bug.cgi?id=181562 <rdar://problem/36445624>
Reviewed by Yusuke Suzuki.

JSTests:

* stress/each-block-at-top-of-polymorphic-call-inlining-should-be-exitOK.js: Added.
(f):
(foo):

Source/JavaScriptCore:

Before this patch, the very first block in the switch for polymorphic call
inlining will have exitOK at the top. The others are not guaranteed to.
That was just a bug. They're all exitOK at the top. This will lead to crashes
in FixupPhase because we won't have a node in a block that has ExitOK, so
when we fixup various type checks, we assert out.

* dfg/DFGByteCodeParser.cpp:
(JSC::DFG::ByteCodeParser::handleInlining):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@226881 268f45cc-cd09-0410-ab3c-d52691b4dbfc