REGRESSION(r226266): [GTK] RELEASE_ASSERT(reservedZoneSize >= minimumReservedZoneSize) in JSC::VM::updateStackLimits https://bugs.webkit.org/show_bug.cgi?id=181438 <rdar://problem/36376724>
Reviewed by Carlos Garcia Campos.

.:

Build JSC as a shared library.

Stop using -fvisibility=hidden. This is a shame, but it is causing problems that I don't
know how to fix.

* Source/cmake/OptionsGTK.cmake:
* Source/cmake/WebKitCompilerFlags.cmake:

Source/JavaScriptCore:

Roll out the functional changes of r226266. We'll keep the minor CMake library type setting
cleanup, but we have to switch back to building JSC only as a shared library, and we have to
get rid of the version script.

* PlatformGTK.cmake:
* javascriptcoregtk-symbols.map: Removed.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@226945 268f45cc-cd09-0410-ab3c-d52691b4dbfc