ASSERTION FAILED: !m_completionHandler in PingHandle::~PingHandle() https://bugs.webkit.org/show_bug.cgi?id=181746 <rdar://problem/36586248>
Reviewed by Chris Dumez.

Source/WebCore:

Call PingHandle::pingLoadComplete() with an error when NSURLConnection queries
whether the ping is able to respond to an authentication request. (Pings do not
respond to authenticate requests.) It will call the completion handler, nullify
the completion handler, and deallocate the PingHandle. Nullifying the completion
handler is necessary to avoid the assertion failure in ~PingHandle().

Test: http/tests/misc/before-unload-load-image.html

* platform/network/PingHandle.h:

LayoutTests:

Adds a test to ensure that loading a protected image from a beforeunload handler
does not cause an assertion failure.

* http/tests/misc/before-unload-load-image-expected.txt: Added.
* http/tests/misc/before-unload-load-image.html: Added.
* http/tests/misc/resources/before-unload-in-subframe-child.html: Added.
* http/tests/misc/resources/protected/protected-image.php: Added.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@227074 268f45cc-cd09-0410-ab3c-d52691b4dbfc