WebViews created via -initSimpleHTMLDocumentWithStyle: don't propagate attachmentElementEnabled state https://bugs.webkit.org/show_bug.cgi?id=181960 <rdar://problem/36431678>
Reviewed by Tim Horton.

Mail's compose view is initialized via the -initSimpleHTMLDocumentWithStyle: codepath. This is intended to be
a fast path for initializing WebViews, which bypasses certain initialization steps present in the regular -init
path, such as calling -_preferencesChanged: to sync RuntimeEnabledFeatures with WebPreferences.

This means that even though we set -[WebPreferences attachmentElementEnabled] to YES for Mail, the value of the
RuntimeEnabledFeature isn't updated to match. To fix this, we explicitly invoke
RuntimeEnabledFeatures::setAttachmentElementEnabled here to make sure the value is in sync with WebPreferences.

* WebView/WebView.mm:
(-[WebView initSimpleHTMLDocumentWithStyle:frame:preferences:groupName:]):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@227371 268f45cc-cd09-0410-ab3c-d52691b4dbfc