WebDriver: several tests in imported/w3c/webdriver/tests/sessions/new_session/merge.py are failing https://bugs.webkit.org/show_bug.cgi?id=180407
Reviewed by Carlos Alberto Lopez Perez.

Source/WebDriver:

This is actually a bug in the spec, see https://github.com/w3c/webdriver/issues/1203. It seems the idea in any
case is that we should check all the firstMatch capabilities before matching them, and return if there's any
error.

* WebDriverService.cpp:
(WebDriver::WebDriverService::mergeCapabilities const):
(WebDriver::WebDriverService::processCapabilities const):

WebDriverTests:

Remove expectations for tests that are passing now and update the bug number of the other ones referencing
#180407.

* TestExpectations.json:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@227414 268f45cc-cd09-0410-ab3c-d52691b4dbfc