UI process sometimes crashes under -[WKContentView _lookupForWebView:] https://bugs.webkit.org/show_bug.cgi?id=182460 <rdar://problem/33260602>
Reviewed by Wenson Hsieh.

* UIProcess/ios/WKContentViewInteraction.mm:
(-[WKContentView _lookupForWebView:]):
If you have a range selection, but no rects for the selection, retrieving
the 0th element of selectionRects will crash the UI process. To fix, in
this case, use the rect for the starting caret instead.

It doesn't seem like the presentationRect is actually currently used for
the Lookup service, so the only impact is that we shouldn't crash anymore.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@228050 268f45cc-cd09-0410-ab3c-d52691b4dbfc